+++ NeteaseCloudMusic/Base/NavigationController/NCNavigationController.m
@@ -31,13 +31,6 @@
     self.interactivePopGestureRecognizer.delegate = self;
 }
 
-//- (void)pushViewController:(UIViewController *)viewController animated:(BOOL)animated {
-//    if (self.childViewControllers.count > 0) {
-//        viewController.hidesBottomBarWhenPushed = YES;
-//    }
-//    [super pushViewController:viewController animated:animated];
-//}
-
 #pragma mark - Lazy Load
 
 - (UIPanGestureRecognizer *)panGestureRecognizer {
