+++ NeteaseCloudMusic/Base/NavigationController/NCNavigationControllerPopAnimatedTransitionTopToBottom.h
