+++ NeteaseCloudMusic/NCMiniPlayerView.m
@@ -17,8 +17,8 @@
 - (instancetype)init {
     self = [super init];
     if (self) {
-        self.backgroundColor = [UIColor systemTealColor];
-        self.layer.cornerRadius = MINI_PLAYVIEW_HEIGHT * 0.5;
+        self.backgroundColor = [UIColor systemGrayColor];
+//        self.layer.cornerRadius = MINI_PLAYVIEW_HEIGHT * 0.5;
         self.layer.masksToBounds = YES;
     }
     return self;
