+++ NeteaseCloudMusic/Supporting Files/Assets.xcassets/AppIcon.appiconset/AppIcon40x40@3x~iphone.png
