+++ NeteaseCloudMusic/Supporting Files/Assets.xcassets/Contents.json
