+++ NeteaseCloudMusic/Supporting Files/Assets.xcassets/bottomLogo_2200x16.imageset/Contents.json
