+++ NeteaseCloudMusic/Supporting Files/Assets.xcassets/mine.imageset/Contents.json
